News_Analysis/pyproject.toml
[tool.poetry]
name = "news-analysis"
version = "0.1.0"
description = ""
authors = ["zoelinsg <zoelin.sg@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.6"
djangorestframework = "^3.15.2"
django-filter = "^25.1"
requests = "^2.32.3"
pandas = "^2.2.3"
openpyxl = "^3.1.5"
django-cors-headers = "^4.7.0"
pyodbc = "^5.2.0"
matplotlib = "^3.10.0"
seaborn = "^0.13.2"
python-dotenv = "^1.0.1"
textblob = "^0.19.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
